Rename a few of the new files
