+++ src/constraints.ts
@@ -15,7 +15,7 @@ import * as tfc from '@tensorflow/tfjs-core';
 import {doc, serialization, Tensor, tidy} from '@tensorflow/tfjs-core';
 
 import {epsilon} from './backend/common';
-import {getScalar} from './engine/unique_id';
+import {getScalar} from './backend/state';
 // import * as K from './backend/tfjs_backend';
 import {deserializeKerasObject, serializeKerasObject} from './utils/generic_utils';
 // tslint:enable:max-line-length
